refactored mailables and updated configuration for adjustable mail logo for laravel 10x

+++ migrations/2024_01_30_164300_add_expires_at_column_to_personal_access_tokens_table.php
@@ -13,9 +13,13 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::table('personal_access_tokens', function (Blueprint $table) {
-            $table->timestamp('expires_at')->nullable()->after('last_used_at');
-        });
+        if (Schema::hasTable('personal_access_tokens')) {
+            Schema::table('personal_access_tokens', function (Blueprint $table) {
+                if (!Schema::hasColumn('personal_access_tokens', 'expires_at')) {
+                    $table->timestamp('expires_at')->nullable()->after('last_used_at');
+                }
+            });
+        }
     }
 
     /**
@@ -25,8 +29,12 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::table('personal_access_tokens', function (Blueprint $table) {
-            $table->dropColumn('expires_at');
-        });
+        if (Schema::hasTable('personal_access_tokens')) {
+            Schema::table('personal_access_tokens', function (Blueprint $table) {
+                if (Schema::hasColumn('personal_access_tokens', 'expires_at')) {
+                    $table->dropColumn('expires_at');
+                }
+            });
+        }
     }
 };
